chore(pallets/communities): split governance functionalities
