+++ pallets/communities/src/functions/governance/mod.rs
@@ -1,0 +1,4 @@
+use super::*;
+
+mod tally;
+mod tracks;
