+++ pallets/communities/src/tests/mod.rs
@@ -12,7 +12,6 @@ const COMMUNITY: CommunityId = 1;
 const COMMUNITY_ADMIN: AccountId = 42;
 
 mod membership;
-mod mock_poll;
 mod registry;
 
 fn setup() {
